# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Stock Microservice",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://95.183.53.25:8001"
        }
    ],
    "paths": {
        "/api/marketplace/history": {
            "get": {
                "tags": [
                    "Stock"
                ],
                "summary": "Historial de compras",
                "operationId": "43c04a9d153a6b6279684b0063268151",
                "responses": {
                    "200": {
                        "description": "Listado de las compras solicitadas al marketplace",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "type": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "quantity_sold": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "date": {
                                                        "type": "string",
                                                        "example": "2022-06-11T20:30:00.000000Z"
                                                    },
                                                    "product": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "number",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Tomato"
                                                            },
                                                            "available_quantity": {
                                                                "type": "number",
                                                                "example": 5
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/orders/get-order": {
            "post": {
                "tags": [
                    "Stock"
                ],
                "summary": "Crear una orden de entrega",
                "operationId": "83fe39067fedf6d72687cbbc05928c98",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "order_code",
                                    "products"
                                ],
                                "properties": {
                                    "order_code": {
                                        "type": "string",
                                        "example": "2302faca-7f66-4078-86d4-abb0ab54b675"
                                    },
                                    "products": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "product"
                                            ],
                                            "properties": {
                                                "product": {
                                                    "type": "string",
                                                    "example": "Tomato"
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Orden actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "type": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Order created successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "code": {
                                                    "type": "string",
                                                    "example": "2302faca-7f66-4078-86d4-abb0ab54b675"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}